Update Practice Code- Sudoku
FINALY, got the requests import to work from the main python tutorial that I am using. Realized that when I used to the 'Recommended" interpreter it allowed pygame to work but not request, but when I changed the interpreter to 'Window' it only fixed request. After more trouble shooting, I found that using the 'Global' interpreter allowed me to use both imports