Update ProjectRoadMap
Had plenty of errors to adjust in this portion of my project. The formatting for python I still get incorrect especially when it is a function I have not used before, or it comes to indents. I try to code how I write and it is not always the case.